third_party/WebKit/LayoutTests/fast/css/invalidation/lang-pseudo-expected.txt

Use descendant invalidation set for :lang pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(norwegian).backgroundColor is transparent
PASS getComputedStyle(nospan).backgroundColor is transparent
PASS getComputedStyle(english).backgroundColor is transparent
PASS getComputedStyle(enspan).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(norwegian).backgroundColor is green
PASS getComputedStyle(nospan).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(english).backgroundColor is green
PASS getComputedStyle(enspan).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
